Move kwiver applets to arrows.
