+++ arrows/core/applets/CMakeLists.txt
@@ -5,22 +5,18 @@
 set( sources
   register_applets.cxx
 
+  dump_klv.cxx
   render_mesh.cxx
   )
 
 set( headers
+  dump_klv.h
   render_mesh.h
   )
 
 ###
 # Add our plugin path to the default list
 include_directories( ${CMAKE_CURRENT_BINARY_DIR} )
-
-###
-# Add path to applets so plugin loader can find them
-kwiver_make_module_path( ${KWIVER_BINARY_DIR} ${kwiver_plugin_applets} )
-kwiver_add_module_path( ${kwiver_module_path_result} )
-
 
 ###
 # Add applet plugin
@@ -31,4 +27,3 @@ kwiver_add_plugin( kwiver_algo_core_applets
                kwiver_tools_applet
                kwiversys
 )
-
