+++ vital/applets/kwiver_applet.cxx
@@ -47,6 +47,16 @@ kwiver_applet::
 {
 }
 
+
+// ----------------------------------------------------------------------------
+void
+kwiver_applet::
+add_command_options()
+{
+  skip_parse_args();
+}
+
+
 // ----------------------------------------------------------------------------
 cxxopts::ParseResult&
 kwiver_applet::
